refactor: rename to header-menu
